+++ packages/spot-consumer/src/consumers/index.ts
@@ -1,0 +1,2 @@
+export * from './SubaccountConsumer'
+export * from './SpotMarketConsumer'