+++ packages/spot-consumer/src/types/index.ts
@@ -1,0 +1,9 @@
+import {
+  PriceLevel as GrpcPriceLevel,
+  CosmosCoin as GrpcCosmosCoin,
+} from '@injectivelabs/exchange-api/injective_spot_markets_rpc_pb'
+
+export * from './spot-market'
+export * from './subaccount'
+
+export { GrpcPriceLevel, GrpcCosmosCoin }